refactor: deleteMember에서 error 검출 추가

+++ src/main/java/com/example/sinabro/service/AdminService.java
@@ -71,6 +71,9 @@ public class AdminService {
     @Transactional
     public void deleteMember(String name) {
         Users users = usersRepository.findByName(name).orElseThrow(UserNotFoundException::new);
+        if(users.getName().equals(name)) {
+            throw new UserAlreadyUsedException();
+        }
         usersRepository.delete(users);
     }
